Split test_pebble.py (3389 LoC) into separate files
